feat:stock section complete

+++ database/migrations/2020_03_31_000010_add_relationship_fields_to_users_table.php
@@ -9,8 +9,8 @@ class AddRelationshipFieldsToUsersTable extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->unsignedInteger('team_id')->nullable();
-            // $table->foreign('team_id', 'team_fk_1230947')->references('id')->on('teams');
+            $table->unsignedBigInteger('team_id')->nullable();
+            // $table->foreign('team_id')->references('id')->on('teams');
         });
 
     }
